+++ gwt-output-qa/src/main/java/arez/gwt/qa/ArezBuildAsserts.java
@@ -194,7 +194,13 @@ public final class ArezBuildAsserts
     index.assertSymbol( "arez\\.Arez", "currentZone", enabled );
     index.assertSymbol( ".*\\.Arez_.*", "$$arezi$$_context", enabled );
     index.assertSymbol( "arez\\.Node", "_context", enabled );
+    index.assertSymbol( "arez\\.Component", "_context", enabled );
+    index.assertSymbol( "arez\\.Node", "_context", enabled );
     index.assertSymbol( "arez\\.ReactionScheduler", "_context", enabled );
+    index.assertSymbol( "arez\\.SchedulerLock", "_context", enabled );
+    index.assertSymbol( "arez\\.SpyImpl", "_context", enabled );
+    index.assertSymbol( "arez\\.Transaction", "_context", enabled );
+    index.assertSymbol( "arez\\.component\\.MemoizeCache", "_context", enabled );
   }
 
   /**